# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  core.Author:
    properties:
      contributors:
        description: |-
          Additional contributors who helped develop the project.

          Use this to acknowledge team members, collaborators, or external contributors.
        items:
          $ref: '#/definitions/core.Contributor'
        type: array
      email:
        description: |-
          Email contact for project-related communication.

          Use team emails for shared ownership: `"ai-team@company.com"`
        type: string
      name:
        description: |-
          Name of the author or team responsible for the project.

          Examples: `"Jane Smith"`, `"AI Platform Team"`, `"Data Science Division"`
        type: string
      organization:
        description: |-
          Organization or company affiliation.

          Examples: `"ACME Corporation"`, `"AI Research Lab"`, `"Engineering Division"`
        type: string
      url:
        description: |-
          URL to author's profile, repository, or team page.

          Examples: `"https://github.com/username"`, `"https://company.com/team/ai"`
        type: string
    type: object
...
  core.Contributor:
    properties:
      email:
        description: Email address for contributor contact.
        type: string
      name:
        description: Full name of the contributor.
        type: string
      organization:
        description: Organization or team the contributor belongs to.
        type: string
      url:
        description: |-
          URL to contributor's profile or portfolio.

          Examples: `"https://github.com/username"`, `"https://linkedin.com/in/name"`
        type: string
    type: object
...
  projectrouter.ProjectDTO:
    properties:
      author:
        $ref: '#/definitions/core.Author'
      autoload:
        additionalProperties: {}
        type: object
      config:
        additionalProperties: {}
        type: object
      description:
        type: string
      memories:
        items:
          additionalProperties: {}
          type: object
        type: array
      models:
        items:
          additionalProperties: {}
          type: object
        type: array
      monitoring:
        additionalProperties: {}
        type: object
      name:
        type: string
      runtime:
        additionalProperties: {}
        type: object
      schemas:
        items:
          additionalProperties: {}
          type: object
        type: array
      tools:
        items:
          additionalProperties: {}
          type: object
        type: array
      version:
        type: string
      workflows:
        items:
          additionalProperties: {}
          type: object
        type: array
    type: object
...
  router.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  router.ProblemDocument:
    properties:
      code:
        example: invalid_cursor
        type: string
      detail:
        example: Invalid cursor parameter
        type: string
      instance:
        example: /api/v0/workflows
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  router.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/router.ErrorInfo'
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact:
    email: support@compozy.com
    name: Compozy Support
    url: https://github.com/compozy/compozy
  description: Compozy is a Next-level Agentic Orchestration Platform, tasks, and tools. Authentication
    is configurable via the SERVER_AUTH_ENABLED environment variable. When enabled, all endpoints require
    Bearer token authentication. When disabled (typical for development), endpoints are accessible without
    authentication.
  license:
    name: BSL-1.1
    url: https://github.com/compozy/compozy/blob/main/LICENSE
  termsOfService: https://github.com/compozy/compozy
  title: Compozy API
  version: "1.0"
paths:
...
  /project:
    delete:
      description: Project deletion is not supported; returns 405.
      produces:
        - application/json
      responses:
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Delete project
      tags:
        - project
    get:
      consumes:
        - application/json
      description: Retrieve the active project configuration.
      parameters:
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Project retrieved
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/projectrouter.ProjectDTO'
                type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Get project
      tags:
        - project
    put:
      consumes:
        - application/json
      description: Create the project configuration when absent or update an existing one using strong
        ETag concurrency.
      parameters:
        - description: Project override
          example: '"demo"'
          in: query
          name: project
          type: string
        - description: Strong ETag for optimistic concurrency
          example: '"\"abc123\""'
          in: header
          name: If-Match
          type: string
        - description: Project configuration payload
          in: body
          name: payload
          required: true
          schema:
            additionalProperties: true
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: Project updated
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/projectrouter.ProjectDTO'
                type: object
        "201":
          description: Project created
          headers:
            ETag:
              description: Strong entity tag for concurrency control
              type: string
            Location:
              description: Absolute URL for the project
              type: string
            RateLimit-Limit:
              description: Requests allowed in the current window
              type: string
            RateLimit-Remaining:
              description: Remaining requests in the current window
              type: string
            RateLimit-Reset:
              description: Seconds until the window resets
              type: string
          schema:
            allOf:
              - $ref: '#/definitions/router.Response'
              - properties:
                  data:
                    $ref: '#/definitions/projectrouter.ProjectDTO'
                type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "412":
          description: ETag mismatch
          schema:
            $ref: '#/definitions/router.ProblemDocument'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/router.ProblemDocument'
      summary: Create or update project
      tags:
        - project
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication (Optional - controlled by SERVER_AUTH_ENABLED environment
      variable. When enabled, all API endpoints require valid authentication. In development/testing,
      authentication may be disabled.)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...